+++ R/constants.R
@@ -1,7 +1,13 @@
 # Choices to pick from
 colours.available <- c('None' = NA, colors())
 text.faces <- c('plain', 'italic', 'bold', 'bold.italic')
-text.families <- names(pdfFonts())
+
+text.families <- if (is.element('extrafont', installed.packages()[, 1])) {
+  c(c('sans', 'serif', 'mono'), extrafont::fonttable()$FamilyName)
+} else {
+  names(pdfFonts())
+}
+
 legend.positions <- c('none', 'left', 'right', 'top', 'bottom', 'XY')
 legend.directions <- c('horizontal', 'vertical')
 linetypes <- c('blank', 'solid', 'dashed', 'dotted', 'dotdash', 'longdash', 'twodash')
@@ -14,7 +20,7 @@ heading.height <- '30px'
 
 # default values
 default <- list(
-  axis.text = list(
+  axis.text = structure(list(
     family = 'sans',
     size = 10,
     face = 'plain',
@@ -23,8 +29,8 @@ default <- list(
     vjust = 0.5,
     angle = 0,
     lineheight = 1.1
-    ),
-  axis.text.x = list(
+    ), class = 'element_text'),
+  axis.text.x = structure(list(
     family = 'sans',
     size = 10,
     face = 'plain',
@@ -33,8 +39,8 @@ default <- list(
     vjust = 1,
     angle = 0,
     lineheight = 1.1
-  ),
-  axis.text.y = list(
+  ), class = 'element_text'),
+  axis.text.y = structure(list(
     family = 'sans',
     size = 10,
     face = 'plain',
@@ -43,18 +49,18 @@ default <- list(
     vjust = 0.5,
     angle = 0,
     lineheight = 1.1
-  ),
-  axis.line = list(
+  ),  class = 'element_text'),
+  axis.line = structure(list(
     colour = 'black',
     size = 1,
     linetype = 'blank'
-    ),
-  axis.ticks = list(
+    ), class = 'element_line'),
+  axis.ticks = structure(list(
     colour = 'grey20',
     size = 0.5,
     linetype = 'solid'
-  ),
-  axis.title = list(
+  ), class = 'element_line'),
+  axis.title = structure(list(
     family = 'sans',
     size = 11,
     face = 'plain',
@@ -63,8 +69,8 @@ default <- list(
     vjust = 0.5,
     angle = 0,
     lineheight = 1.1
-  ),
-  plot.title = list(
+  ), class = 'element_text'),
+  plot.title = structure(list(
     family = 'sans',
     size = 13,
     face = 'plain',
@@ -73,53 +79,53 @@ default <- list(
     vjust = 0.5,
     angle = 0,
     lineheight = 1.1
-  ),
-  panel.background = list(
+  ), class = 'element_text'),
+  panel.background = structure(list(
     fill = 'grey92',
     colour = 'NA',
     size = 0.5,
     linetype = 'blank'
-  ),
-  plot.background = list(
+  ), class = 'element_rect'),
+  plot.background = structure(list(
     fill = 'NA',
     colour = 'white',
     size = 0.5,
     linetype = 'blank'
-  ),
-  panel.grid.major = list(
+  ), class = 'element_rect'),
+  panel.grid.major = structure(list(
     colour = 'grey100',
     size = 0.5,
     linetype = 'solid'
-  ),
-  panel.grid.minor = list(
+  ), class = 'element_line'),
+  panel.grid.minor = structure(list(
     colour = 'grey100',
     size = 0.5,
     linetype = 'solid'
-  ),
-  legend.text = list(
+  ), class = 'element_line'),
+  legend.text = structure(list(
     size = 10,
     face = 'plain',
     family = 'sans',
     colour = 'black'
-  ),
-  legend.title = list(
+  ), class = 'element_text'),
+  legend.title = structure(list(
     size = 10,
     face = 'plain',
     family = 'sans',
     colour = 'black'
-  ),
-  legend.background = list(
+  ), class = 'element_text'),
+  legend.background = structure(list(
     fill = 'grey100',
     colour = 'NA',
     size = 0.5,
     linetype = 'blank'
-  ),
-  legend.key = list(
+  ), class = 'element_rect'),
+  legend.key = structure(list(
     fill = 'grey95',
     colour = 'NA',
     size = 0.5,
     linetype = 'blank'
-  ),
+  ), class = 'element_rect'),
   legend.position = 'right',
   legend.position.x = 0.5,
   legend.position.y = 0.5,